+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to
 
 ### Fixed
 
+- Fix credential not added automatically after being created from the canvas.
+  [#2105](https://github.com/OpenFn/lightning/issues/2105)
 - Replace the "not working?" prompt by "All good, but if your credential stops
   working, you may need to re-authorize here.".
   [#2102](https://github.com/OpenFn/lightning/issues/1872)